+++ pubspec.yaml
@@ -38,6 +38,7 @@ dependencies:
   firebase_core: ^3.8.0
   cloud_firestore: ^5.5.0
   firebase_auth: ^5.3.3
+  go_router: ^14.6.0
 
 dev_dependencies:
   flutter_test:
